+++ package.json
@@ -3,7 +3,7 @@
   "version": "0.1.0",
   "description": "A TypeScript library for building data nodes from JSON-defined structures and relationships.",
   "main": "dist/index.js",
-  "module": "dist/index.mjs",
+  "module": "dist/index.es.js",
   "types": "dist/index.d.ts",
   "scripts": {
     "build": "vite build",
