+++ src/utils/schema.ts
@@ -1,29 +1,9 @@
 import { BigNumber } from "bignumber.js"
 import { SchemaType } from "../enum/schemaType"
-import { registerSchema } from "./schemaProvider"
+import { registerSchema, NS_SYSTEM_ARRAY, NS_SYSTEM_BOOL, NS_SYSTEM_DATE, NS_SYSTEM_FULLDATE, NS_SYSTEM_INT, NS_SYSTEM_NUMBER, NS_SYSTEM_STRING, NS_SYSTEM_STRUCT, NS_SYSTEM_YEAR, NS_SYSTEM_YEARMONTH, NS_SYSTEM_DOUBLE, NS_SYSTEM_FLOAT, NS_SYSTEM_INTS, NS_SYSTEM_NUMBERS, NS_SYSTEM_RANGEDATE, NS_SYSTEM_RANGEFULLDATE, NS_SYSTEM_RANGEMONTH, NS_SYSTEM_RANGEYEAR, NS_SYSTEM_STRINGS } from "./schemaProvider"
 import { _LS } from "./locale"
 import { isNull } from "./toolset"
 import { ExpressionType } from "../enum/expressionType"
-
-export const NS_SYSTEM_ARRAY = "system.array"
-export const NS_SYSTEM_STRUCT = "system.struct"
-export const NS_SYSTEM_BOOL = "system.bool"
-export const NS_SYSTEM_DATE = "system.date"
-export const NS_SYSTEM_NUMBER = "system.number"
-export const NS_SYSTEM_DOUBLE = "system.double"
-export const NS_SYSTEM_FLOAT = "system.float"
-export const NS_SYSTEM_FULLDATE = "system.fulldate"
-export const NS_SYSTEM_INT = "system.int"
-export const NS_SYSTEM_STRING = "system.string"
-export const NS_SYSTEM_YEAR = "system.year"
-export const NS_SYSTEM_YEARMONTH = "system.yearmonth"
-export const NS_SYSTEM_RANGEDATE = "system.rangedate"
-export const NS_SYSTEM_RANGEFULLDATE = "system.rangefulldate"
-export const NS_SYSTEM_RANGEMONTH = "system.rangemonth"
-export const NS_SYSTEM_RANGEYEAR = "system.rangeyear"
-export const NS_SYSTEM_STRINGS = "system.strings"
-export const NS_SYSTEM_NUMBERS = "system.numbers"
-export const NS_SYSTEM_INTS = "system.ints"
 
 /**
  * The default schemas
